Program_Remembers_Your_Name
(1) Type your name in the left side of the bar.
(2) You will get a greeting.
(3) Close the program.
(4) Open it up and click the left bar again. It will ask if your name is (name you typed in before you closed the bar).
(5) If it is, it will say, "Welcome back, (your name)" and will show you the time you last logged in.

I used the localstorage syntax for this.